更新图片修改

+++ src/main/java/com/movision/mybatis/systemLayout/mapper/SystemLayoutMapper.java
@@ -21,5 +21,5 @@ public interface SystemLayoutMapper {
 
     String queryServiceUrl(String fileurl);
 
-    Integer queryFileRatio(String ratio);
+    Double queryFileRatio(String ratio);
 }
